refactor: DTO int -> BigDecimal로 변경, Stirng -> ENUM타입으로 변경

+++ src/main/java/com/project/baedalsodae/store/dto/response/store/OwnerStoreResponse.java
@@ -2,6 +2,7 @@ package com.project.baedalsodae.store.dto.response.store;
 
 import com.project.baedalsodae.global.common.dto.AddressResponse;
 import com.project.baedalsodae.store.entity.Store;
+import java.math.BigDecimal;
 import java.util.UUID;
 import lombok.Builder;
 import lombok.Getter;
@@ -14,7 +15,7 @@ public class OwnerStoreResponse {
     private String phoneNumber;
     private String businessNumber;
     private String description;
-    private Integer reviewCount;
+    private BigDecimal reviewCount;
     private Double avgRating;
     private String storeStatus;
     private AddressResponse address;
@@ -26,7 +27,7 @@ public class OwnerStoreResponse {
                 .phoneNumber(store.getPhone())
                 .businessNumber(store.getBusinessNumber())
                 .description(store.getDescription())
-                .reviewCount(store.getReviewCount())
+                .reviewCount(BigDecimal.valueOf(store.getReviewCount()))
                 .avgRating(store.getAvgRating())
                 .storeStatus(store.getStoreStatus().name())
                 .address(AddressResponse.fromEntity(store.getAddress()))
