+++ src/main/java/com/project/baedalsodae/store/dto/response/storeCategory/StoreCategoryListResponse.java
@@ -1,5 +1,6 @@
 package com.project.baedalsodae.store.dto.response.storeCategory;
 
+import java.math.BigDecimal;
 import java.util.List;
 import lombok.*;
 
@@ -7,13 +8,13 @@ import lombok.*;
 @Builder
 public class StoreCategoryListResponse {
     private List<StoreCategoryResponse> storeCategoryList;
-    private int totalCount;
+    private BigDecimal totalCount;
 
     public static StoreCategoryListResponse fromList(
             List<StoreCategoryResponse> storeCategoryList) {
         return StoreCategoryListResponse.builder()
                 .storeCategoryList(storeCategoryList)
-                .totalCount(storeCategoryList.size())
+                .totalCount(BigDecimal.valueOf(storeCategoryList.size()))
                 .build();
     }
 }
